refactor(events): move/rename utils/formats -> utils/format and add components barrel
