+++ frontend/src/features/events/components/index.js
@@ -1,0 +1,6 @@
+export { default as EventsFilters } from './EventsFilters';
+export { default as EventsBoard } from './EventsBoard';
+export { default as EventModal } from './EventModal';
+export { default as AttendanceModalContent } from './AttendanceModal';
+export { default as EventsForm } from './EventsForm';
+export { default as EventsSummaryCards } from './EventsSummaryCards';
